Refactor lowering for expressions
