+++ lang/lowering/src/lower/mod.rs
@@ -6,13 +6,13 @@ mod exp;
 pub trait Lower {
     type Target;
 
-    fn lower(&self, ctx: &mut Ctx) -> Result<Self::Target, LoweringError>;
+    fn lower(&self, ctx: &mut Ctx) -> LoweringResult<Self::Target>;
 }
 
 impl<T: Lower> Lower for Option<T> {
     type Target = Option<T::Target>;
 
-    fn lower(&self, ctx: &mut Ctx) -> Result<Self::Target, LoweringError> {
+    fn lower(&self, ctx: &mut Ctx) -> LoweringResult<Self::Target> {
         self.as_ref().map(|x| x.lower(ctx)).transpose()
     }
 }
@@ -20,7 +20,7 @@ impl<T: Lower> Lower for Option<T> {
 impl<T: Lower> Lower for Vec<T> {
     type Target = Vec<T::Target>;
 
-    fn lower(&self, ctx: &mut Ctx) -> Result<Self::Target, LoweringError> {
+    fn lower(&self, ctx: &mut Ctx) -> LoweringResult<Self::Target> {
         self.iter().map(|x| x.lower(ctx)).collect()
     }
 }
@@ -28,7 +28,7 @@ impl<T: Lower> Lower for Vec<T> {
 impl<T: Lower> Lower for Box<T> {
     type Target = Box<T::Target>;
 
-    fn lower(&self, ctx: &mut Ctx) -> Result<Self::Target, LoweringError> {
+    fn lower(&self, ctx: &mut Ctx) -> LoweringResult<Self::Target> {
         Ok(Box::new((**self).lower(ctx)?))
     }
 }
