+++ lang/lowering/src/result.rs
@@ -2,6 +2,10 @@ use miette::{Diagnostic, SourceSpan};
 use parser::cst::ident::Ident;
 use thiserror::Error;
 
+/// The result type specialized to lowering errors.
+pub type LoweringResult<T = ()> = Result<T, LoweringError>;
+
+/// All the errors that can be emitted during lowering
 #[derive(Error, Diagnostic, Debug, Clone)]
 pub enum LoweringError {
     #[error("Undefined identifier {}", name.id)]
